+++ src/scripts/train/OLMo3/OLMo-3-190M.py
@@ -70,11 +70,11 @@ def build_experiment_config(cli_context: CliContext) -> ExperimentConfig:
 
     tokenizer_config = TokenizerConfig.dolma2()
 
+    attn_backend = (
+        AttentionBackendName.flash_2 if "B200" in gpu_type else AttentionBackendName.flash_3
+    )
     model_config = TransformerConfig.olmo3_190M(
-        vocab_size=tokenizer_config.padded_vocab_size(),
-        attn_backend=AttentionBackendName.flash_2
-        if "B200" in gpu_type
-        else AttentionBackendName.flash_3,
+        vocab_size=tokenizer_config.padded_vocab_size(), attn_backend=attn_backend
     )
 
     train_module_config = TransformerTrainModuleConfig(